Refactor - move PackageInfo.Platform graphPlatform, destinations to public extension
