+++ Sources/TuistDependencies/SwiftPackageManager/Utils/PackageInfoMapper.swift
@@ -1076,44 +1076,6 @@ extension ProjectDescription.Settings {
     }
 }
 
-extension PackageInfo.Platform {
-    fileprivate func graphPlatform() throws -> TuistGraph.Platform {
-        switch platformName.lowercased() {
-        case "ios", "maccatalyst":
-            return .iOS
-        case "macos":
-            return .macOS
-        case "tvos":
-            return .tvOS
-        case "watchos":
-            return .watchOS
-        case "visionos":
-            return .visionOS
-        default:
-            throw PackageInfoMapperError.unknownPlatform(platformName)
-        }
-    }
-
-    fileprivate func destinations() throws -> ProjectDescription.Destinations {
-        switch platformName.lowercased() {
-        case "ios":
-            return [.iPhone, .iPad, .macWithiPadDesign, .appleVisionWithiPadDesign]
-        case "maccatalyst":
-            return [.macCatalyst]
-        case "macos":
-            return [.mac]
-        case "tvos":
-            return [.appleTv]
-        case "watchos":
-            return [.appleWatch]
-        case "visionos":
-            return [.appleVision]
-        default:
-            throw PackageInfoMapperError.unknownPlatform(platformName)
-        }
-    }
-}
-
 extension ProjectDescription.Product {
     fileprivate static func from(product: TuistGraph.Product) -> Self {
         switch product {
